requirements.txt
fastapi
uvicorn
openai
python-dotenv
pinecone
streamlit
requests
beautifulsoup4
tqdm